Don't re-export files that already exist (allows for fast bugfixing the HTML generation)

+++ MediaExportHelper.cs
@@ -7,11 +7,22 @@ public static class MediaExportHelper
     public static void CopyMediaAndEmitHtml(string contentPath, string? mimeType, string outputMediaRelativePath, string outputMediaAbsolutePath,
         StringBuilder sb)
     {
+        if (string.IsNullOrEmpty(mimeType)) // guess based on file extension
+        {
+            if (outputMediaAbsolutePath.EndsWith(".jpg", StringComparison.OrdinalIgnoreCase) ||
+                outputMediaAbsolutePath.EndsWith(".jpeg", StringComparison.OrdinalIgnoreCase)) mimeType = "image/jpeg";
+            else if (outputMediaAbsolutePath.EndsWith(".png", StringComparison.OrdinalIgnoreCase)) mimeType = "image/png";
+            else if (outputMediaAbsolutePath.EndsWith(".mp4", StringComparison.OrdinalIgnoreCase)) mimeType = "video/mp4";
+            else if (outputMediaAbsolutePath.EndsWith(".mov", StringComparison.OrdinalIgnoreCase)) mimeType = "video/quicktime";
+        }
+
         switch (mimeType)
         {
             case "video/mp4":
             case "video/quicktime":
             case "video/3gpp":
+                if (File.Exists(outputMediaAbsolutePath)) break; // resume of a previous export, the file already exists 
+
                 Console.WriteLine("Copying {0} to {1}", mimeType, outputMediaRelativePath);
                 File.Copy(contentPath, outputMediaAbsolutePath, overwrite: true);
 
@@ -35,8 +46,11 @@ public static class MediaExportHelper
                     var newRelative = Util.ChangeFileExtension(outputMediaRelativePath, ".avif");
                     var newAbsolute = Util.ChangeFileExtension(outputMediaAbsolutePath, ".avif");
 
-                    Console.WriteLine("Converting {0} to {1}", mimeType, newRelative);
-                    ImageConverter.JpegOrPngToAvif(contentPath, newAbsolute);
+                    if (!File.Exists(newAbsolute))
+                    {
+                        Console.WriteLine("Converting {0} to {1}", mimeType, newRelative);
+                        ImageConverter.JpegOrPngToAvif(contentPath, newAbsolute);
+                    }
 
                     // assume it's an image if we can't tell what else it might have been
                     sb.AppendLine($"<img class=\"image-attachment\" src=\"{newRelative}\" />");
@@ -54,8 +68,11 @@ public static class MediaExportHelper
                     var newRelative = Util.ChangeFileExtension(outputMediaRelativePath, ".avif");
                     var newAbsolute = Util.ChangeFileExtension(outputMediaAbsolutePath, ".avif");
 
-                    Console.WriteLine("Converting JPEG to {0}", newRelative);
-                    ImageConverter.HeicToAvif(contentPath, newAbsolute);
+                    if (!File.Exists(newAbsolute))
+                    {
+                        Console.WriteLine("Converting JPEG to {0}", newRelative);
+                        ImageConverter.HeicToAvif(contentPath, newAbsolute);
+                    }
 
                     // assume it's an image if we can't tell what else it might have been
                     sb.AppendLine($"<img class=\"image-attachment\" src=\"{newRelative}\" />");
@@ -68,8 +85,11 @@ public static class MediaExportHelper
                 }
 
             default:
-                Console.WriteLine("Copying to {0}", outputMediaRelativePath);
-                File.Copy(contentPath, outputMediaAbsolutePath, overwrite: true);
+                if (!File.Exists(outputMediaAbsolutePath))
+                {
+                    Console.WriteLine("Copying to {0}", outputMediaRelativePath);
+                    File.Copy(contentPath, outputMediaAbsolutePath, overwrite: true);
+                }
                 // assume it's an image if we can't tell what else it might have been
                 sb.AppendLine($"<img class=\"image-attachment\" src=\"{outputMediaRelativePath}\" />");
                 break;
